feat(ci): optimize build process

+++ docker-entrypoint.sh
@@ -1,12 +1,17 @@
 #!/bin/sh
 
+# If we already have a config file, simply drop setup folder
+if [ -f "/var/www/html/include/ost-config.php" ]; then
+    rm -rf /var/www/html/setup
+fi
+
 # Create database etc. if it does not exist yet, but make sure to use our own configuration
-if [ -d "/var/www/html/setup/" ]; then
+if [ -d "/var/www/html/setup/" -a ! -f "/var/www/html/include/ost-config.php" ]; then
     cp -rf /config/* /var/www/html/include
     chmod +w /var/www/html/include/ost-config.php
     cd /var/www/html/setup
     php ./envbasedinstall.php
-    if [ $? -eq 0 ]; then
+    if [ -f "/var/www/html/setup/success" ]; then
         cd ..
         rm -rf /var/www/html/setup
         chmod 0644 /var/www/html/include/ost-config.php
